// packages/game-engine/src/types.ts
export enum TeamType {
  VILLAGE = 'VILLAGE',
  WEREWOLF = 'WEREWOLF',
  NEUTRAL = 'NEUTRAL',
}

export enum RoleType {
  VILLAGER = 'VILLAGER',
  WEREWOLF = 'WEREWOLF',
  ALPHA = 'ALPHA',
  BETA = 'BETA',
  SEER = 'SEER',
  WITCH = 'WITCH',
  GUARDIAN = 'GUARDIAN',
  HUNTER = 'HUNTER',
  ELDER = 'ELDER',
  PRIEST = 'PRIEST',
  ORACLE = 'ORACLE',
  INQUISITOR = 'INQUISITOR',
  SHAMAN = 'SHAMAN',
  JESTER = 'JESTER',
  TRAITOR = 'TRAITOR',
  SHAPESHIFTER = 'SHAPESHIFTER',
  SHADOW = 'SHADOW',
  HOWLER = 'HOWLER',
  DEVOURER = 'DEVOURER',
  CURSED = 'CURSED',
  BLOOD_MOON = 'BLOOD_MOON',
  DARK_SHAMAN = 'DARK_SHAMAN',
  WOLF_KING = 'WOLF_KING',
  ANCESTRAL_DRUID = 'ANCESTRAL_DRUID',
  NECROMANCER = 'NECROMANCER',
  PHOENIX = 'PHOENIX',
  JUDGE = 'JUDGE',
  RUNE_GUARDIAN = 'RUNE_GUARDIAN',
}

export enum PhaseType {
  SETUP = 'SETUP',
  NIGHT = 'NIGHT',
  DAWN = 'DAWN',
  DISCUSSION = 'DISCUSSION',
  VOTING = 'VOTING',
  RESOLUTION = 'RESOLUTION',
  GAME_OVER = 'GAME_OVER',
}

export enum ActionType {
  BLOCK = 'BLOCK',
  PROTECT = 'PROTECT',
  INVESTIGATE = 'INVESTIGATE',
  TRACK = 'TRACK',
  MARK = 'MARK',
  CONVERT = 'CONVERT',
  ATTACK = 'ATTACK',
  EXECUTE = 'EXECUTE',
  REVIVE = 'REVIVE',
  VOTE = 'VOTE',
  SPECIAL = 'SPECIAL',
}

export enum EventType {
  GAME_CREATED = 'GAME_CREATED',
  ROLE_ASSIGNED = 'ROLE_ASSIGNED',
  NIGHT_STARTED = 'NIGHT_STARTED',
  PLAYER_ATTACKED = 'PLAYER_ATTACKED',
  PLAYER_KILLED = 'PLAYER_KILLED',
  PLAYER_PROTECTED = 'PLAYER_PROTECTED',
  PHASE_STARTED = 'PHASE_STARTED',
  PHASE_ENDED = 'PHASE_ENDED',
  DAY_STARTED = 'DAY_STARTED',
  VOTE_CAST = 'VOTE_CAST',
  PLAYER_EXECUTED = 'PLAYER_EXECUTED',
  VICTORY_EVALUATED = 'VICTORY_EVALUATED',
  GAME_ENDED = 'GAME_ENDED',
}